Fixed class naming, divided tests
